+++ Fun.Build/ProcessExtensions.fs
@@ -1,6 +1,8 @@
 ﻿[<AutoOpen>]
 module Fun.Build.ProcessExtensions
 
+#nowarn "9"
+
 open System
 open System.IO
 open System.Text
@@ -9,6 +11,66 @@ open System.Diagnostics
 open System.Runtime.InteropServices
 open Spectre.Console
 open Fun.Build.Internal
+
+module private Native =
+
+    module private Windows =
+
+        let terminate (p: Process) =
+            try
+                if p.CloseMainWindow() then
+                    Ok p.ExitCode
+                else
+                    Error "Process has no main window or the main window is disabled"
+            with ex ->
+                Error ex.Message
+
+    module private Unix =
+        open Microsoft.FSharp.NativeInterop
+
+        [<DllImport("libc", SetLastError = true)>]
+        extern int private kill(int pid, int signal)
+
+        [<DllImport("libc", SetLastError = true)>]
+        extern int private strerror_r(int errnum, char* buf, UInt64 buflen)
+
+        let private getErrorMessage errno =
+            let buffer = NativePtr.stackalloc<char> 1024
+            let result = strerror_r (errno, buffer, 1024UL)
+
+            if result = 0 then
+                Marshal.PtrToStringAnsi(buffer |> NativePtr.toNativeInt)
+            else
+                $"errno %i{errno}"
+
+        [<Literal>]
+        let private SIGTERM = 15
+
+        [<Literal>]
+        let private ESRCH = 3
+
+        let terminate (p: Process) =
+            try
+                let code = kill (p.Id, SIGTERM)
+                let errno = Marshal.GetLastWin32Error()
+                if code = -1 && errno <> ESRCH then // ESRCH = process does not exist, assume it exited
+                    getErrorMessage errno |> Error
+                else
+                    Ok code
+            with ex ->
+                Error ex.Message
+
+    let kill (p: Process) =
+        let result =
+            if RuntimeInformation.IsOSPlatform OSPlatform.Windows then
+                Windows.terminate p
+            else
+                Unix.terminate p
+
+        match result with
+        | Ok _ -> ()
+        | Error _ -> p.Kill()
+
 
 type Process with
 
@@ -86,10 +148,8 @@ type Process with
             use! cd =
                 Async.OnCancel(fun _ ->
                     AnsiConsole.Markup $"[yellow]{logPrefix}[/] "
-
-                    AnsiConsole.WriteLine $"{commandLogString} is cancelled or timed out and the process will be killed."
-
-                    result.Kill()
+                    AnsiConsole.WriteLine $"{commandLogString}: is cancelled or timed out and the process will be killed."
+                    Native.kill result
                 )
 
             use _ =
@@ -97,7 +157,7 @@ type Process with
                 | Some ct ->
                     ct.Register(fun () ->
                         AnsiConsole.MarkupLine("[yellow]Command is cancelled by your token[/]")
-                        result.Kill()
+                        Native.kill result
                     )
                     :> IDisposable
                 | _ ->